Progression + Fix

+++ protocol/protocol_primitives.h
@@ -30,7 +30,7 @@ typedef enum content_type {
 } content_type;
 
 typedef union content_union {
-	long int integer;
+	int integer;
 	char* string;
 } content_union;
 
@@ -45,7 +45,7 @@ typedef struct data_element {
 } data_element;
 
 typedef struct protocol_data {
-	long int total_length;
+	int total_length;
 	message_type type;
 	data_element* data;
 } protocol_data;
@@ -66,9 +66,9 @@ void insertData(protocol_data* d, content_data* content);
 
 const char* encodeType(message_type type);
 
-char* encodeNumber(long int n, int length);
+char* encodeNumber(int n, int length);
 
-char* encodeString(char* string, long int length);
+char* encodeString(char* string, int length);
 
 protocol_data* initMessageHeader(message_type type);
 
@@ -76,7 +76,7 @@ protocol_message encodeProtocolData(protocol_data* d);
 
 void addMessageString(protocol_data* d, char* string);
 
-void addMessageNumber(protocol_data* d, long int number);
+void addMessageNumber(protocol_data* d, int number);
 
 /* ---------------------------------------------
 			FONCTIONS DE DECODAGE
@@ -84,9 +84,9 @@ void addMessageNumber(protocol_data* d, long int number);
 
 int char2int(char c);
 
-long int decodeNumber(char* number);
+int decodeNumber(char* number);
 
-long int decodeLength(protocol_message message);
+int decodeLength(protocol_message message);
 
 // Retourn -1 si échec
 message_type decodeType(protocol_message message);
